Se separan los tests de las páginas de error.
